Little bug fixed by Cameron (cam@kego.com)
git-svn-id: https://svn.ic-s.nl/svn/dbmail/trunk/dbmail@364 7b491191-dbf0-0310-aff6-d879d4d69008

+++ sql/create_tables.sql
@@ -46,7 +46,7 @@ CREATE TABLE mailbox (
   
 CREATE TABLE message (
    messageidnr bigint(21) DEFAULT '0' NOT NULL auto_increment,
-   mailboxidnr int(21) DEFAULT '0' NOT NULL,
+   mailboxidnr bigint(21) DEFAULT '0' NOT NULL,
    messagesize bigint(21) DEFAULT '0' NOT NULL,
    seen_flag tinyint(1) default '0' not null,
    answered_flag tinyint(1) default '0' not null,
